+++ CHANGELOG.md
@@ -7,7 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [1.3.0]
 
-- added search to `khao-icon`
+- added icons `navigate-back` and `navigate-forward` to to `khao-icon`
+- added `search` icon to `khao-icon`
 - changed help icon in `khao-icon`
 - added `opensNewWindow` attribute to `khao-link`
 - (!) ommited `khao-card-notice`, you can use `khao-card` to layout card accordingly